Add files via upload
The objective of this project is to model the degradation of the TileCal scintillators and optical fibres with Machine Learning regression to investigate the dependencies on the specific detector design parameters (scintillators/fibres sizes and others) and run conditions (integrated dose and dose rate). The aim is, for instance, to determine which elements of the detector are the weakest links in terms of damage: the optical fibres collecting light from the scintillators or the scintillator tiles themselves. Such results would add to a current fundamental discussion about state-of-the-art calorimetry.